Project views and models were improved.

+++ BilliT/ticket/models.py
@@ -5,7 +5,7 @@ class BusTicket(models.Model):
     src_town = models.CharField(max_length=1024, verbose_name="Source Town")
     dst_town = models.CharField(max_length=1024, verbose_name="Destination Town")
     buyer = models.ForeignKey('user.User')
-    fi = models.DecimalField()
+    fi = models.IntegerField()
     dispatch_time = models.DateTimeField()
 
 
@@ -13,8 +13,8 @@ class AirplaneTicket(models.Model):
     src_town = models.CharField(max_length=1024, verbose_name="Source Town")
     dst_town = models.CharField(max_length=1024, verbose_name="Destination Town")
     buyer = models.ForeignKey('user.User')
-    flight_no = models.DecimalField()
-    fi = models.DecimalField()
+    flight_no = models.IntegerField()
+    fi = models.IntegerField()
     dispatch_time = models.DateTimeField()
 
 
@@ -22,5 +22,5 @@ class TrainTicket(models.Model):
     src_town = models.CharField(max_length=1024, verbose_name="Source Town")
     dst_town = models.CharField(max_length=1024, verbose_name="Destination Town")
     buyer = models.ForeignKey('user.User')
-    fi = models.DecimalField()
+    fi = models.IntegerField()
     dispatch_time = models.DateTimeField()
